§15 Part 1 Teams: migrate TeamService to Application layer
Implements nobodies-collective/Humans#540.

Move TeamService from Humans.Infrastructure/Services/ to
Humans.Application.Services.Teams. All owned-table access now flows
through the new ITeamRepository (Humans.Application.Interfaces.Repositories)
backed by TeamRepository in Humans.Infrastructure.Repositories.Teams,
using IDbContextFactory<HumansDbContext> so the repository stays Singleton
while DbContext instances remain short-lived. Cross-section reads route
through IUserService / IRoleAssignmentService / IShiftManagementService /
ITeamResourceService / IEmailService / ISystemTeamSync via their public
interfaces; 5 cross-domain .Include(...User) chains in the service are gone.

Decorator decision: Option A (no separate CachingTeamService). The existing
10-minute IMemoryCache projection at CacheKeys.ActiveTeams stays in-service,
same precedent as Camps (§15f / §15i Camps entry). A Singleton decorator
can be layered on later without changing the ITeamService surface. Rationale:
the cache is tightly coupled to 10+ write-path mutation sites; cleaving it
into a separate Singleton decorator would double the surface area for
limited benefit at ~500-user scale, and the non-decorator precedent
(User #243, Budget #544, Auth #551, Camps) is well-established.

Cross-domain navs on Team-owned entities are [Obsolete]-marked per §6c:
TeamMember.User, TeamJoinRequest.User, TeamJoinRequest.ReviewedByUser,
TeamRoleAssignment.AssignedByUser, TeamJoinRequestStateHistory.ChangedByUser.
TeamService populates them in-memory via IUserService.GetByIdsAsync (§6b)
before returning the entity graph, so controllers/views/models continue to
read through them under file-wide CS0618 pragmas. The pragmas clear when
the User-entity nav strip lands (tracked in §15i).

Design-rules §15i updated:
- Teams moved from "partial" to fully migrated
- Business-services-in-Infrastructure count: 4 → 3
- Sections with 0 cross-domain .Include() now explicitly includes all
  three Teams services
- Repository count: 26 → 27 (TeamRepository added)

Architecture tests: TeamsArchitectureTests pins §15 invariants
(namespace, no DbContext, no IDbContextFactory, takes ITeamRepository,
Application assembly, no EF reference, TeamRepository is sealed).

Repository tests: TeamRepositoryTests covers the compound-write paths
(AddMemberWithOutbox, MarkMemberLeftWithOutbox, WithdrawRequest,
DeactivateTeam, RevokeAllMemberships) against the in-memory provider.

EF migration check: `dotnet ef migrations add VerifyTeams` produces an
empty Up/Down pair; removed the scaffolded migration.

Build: clean (0 warnings, 0 errors).
Tests: 1484 Application pass (24 new), 116 Domain pass, 32 Integration pass.
