+++ src/Humans.Infrastructure/Data/Configurations/Teams/TeamRoleAssignmentConfiguration.cs
@@ -38,9 +38,11 @@ public class TeamRoleAssignmentConfiguration : IEntityTypeConfiguration<TeamRole
             .HasForeignKey(a => a.TeamMemberId)
             .OnDelete(DeleteBehavior.Restrict);
 
+#pragma warning disable CS0618 // Obsolete cross-domain nav kept so EF FK constraint stays modelled.
         builder.HasOne(a => a.AssignedByUser)
             .WithMany()
             .HasForeignKey(a => a.AssignedByUserId)
             .OnDelete(DeleteBehavior.Restrict);
+#pragma warning restore CS0618
     }
 }
