+++ src/Humans.Web/Controllers/VolController.cs
@@ -1,4 +1,11 @@
 // @e2e: shifts.spec.ts
+// TeamMember.User / TeamJoinRequest.User / TeamRoleAssignment.AssignedByUser
+// are [Obsolete] cross-domain navs (design-rules §6c). The Teams service
+// populates them in-memory (§6b) before returning the entity graph, so the
+// view-model projections here are safe reads — but the compiler still warns
+// and TreatWarningsAsErrors promotes to error. This file-wide disable is
+// cleared when the controller projects directly from service-layer DTOs.
+#pragma warning disable CS0618
 using Humans.Application.Interfaces.Shifts;
 using Humans.Application.Interfaces.Teams;
 using Humans.Domain.Constants;
